Charts/Line_bar_pie.PNG
Please see below the link to my tableau profile. Included is the profile for this project in which shows in greater details of my research findings for the trending jobs. For an example, in tab 3 (salary trends across grades) you can see the highest paid position, with the highest grade & average base salary is an information tech project manager. While an office clerk is the lowest paid position. 

*[Jasmine Grant Tableau Vizzes](https://public.tableau.com/app/profile/jasmine.grant/vizzes)